requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
twilio==8.5.0
anthropic==0.7.0
selenium==4.11.2
webdriver-manager==4.0.0
pydub==0.25.1
SpeechRecognition==3.10.0
schedule==1.2.0
pytest==7.4.0 